[VGNWB-282] Resolve merge conflicts
