+++ src/components/texts/TextAdditionalContent.vue
@@ -26,7 +26,7 @@ import { AxiosResponse } from 'axios';
 import { Component, Prop, Vue } from 'vue-property-decorator';
 
 import EmptyView from '@/components/global/EmptyView.vue';
-import { TextContentResource } from '@/models';
+import { TextContentResource } from '@/interfaces/RepositoryTypes';
 
 @Component({
 	name: 'TextAdditionalContent',
