+++ src/views/texts/RepoList.vue
@@ -27,7 +27,7 @@ import { Component, Vue } from 'vue-property-decorator';
 
 import ErrorIndicator from '@/components/global/ErrorIndicator.vue';
 import Loading from '@/components/global/Loading.vue';
-import { Repository } from '@/models';
+import { Repository } from '@/interfaces/RepositoryTypes';
 
 @Component({
 	name: 'RepoList',
